requirements.txt
boto3==1.4.4
botocore==1.5.9
docutils==0.13.1
jmespath==0.9.1
python-dateutil==2.6.0
PyYAML==3.12
s3transfer==0.1.10
six==1.10.0
